+++ Nokia.AssessmentMange.Domain/Application/ISubjectApplication.cs
@@ -19,7 +19,7 @@ namespace Nokia.AssessmentMange.Domain.Application
         /// <param name="sex"></param>
         /// <param name="ageRange"></param>
         /// <param name="score"></param>
-        ConversionTable InitConversion(string subjectId, Sex sex, AgeRange ageRange, double score);
-        ConversionTable AddScore(string subjectId, Sex sex, double score);
+        ConversionTable InitConversion(string subjectId, Sex sex, AgeRange ageRange, int score);
+        ConversionTable AddScore(string subjectId, Sex sex, int score);
     }
 }
